requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.1
WTForms==2.3.3
